Made the driver use the real BrowserKit API
The HttpKernel client implementation does not respect the BrowserKit
expectation for the getResponse() method by returning the HttpFoundation
response. The BrowserKitDriver was previously written against this broken
implementation, making it incompatible with proper BrowserKit
implemenation.
The driver now converts the HttpFoundation response to a BrowserKit
response to take the HttpKernel implementation into account while still
relying on the proper API.