Deprecated the URL alteration in visit()
This was introduced to strip the base url from the url when using the
HttpKernel client in early versions of the driver. However, this was only
half-working as it was impacting the current url later. A proper way to
support the base url has been introduced for the HttpKernel client, so the
old one can be deprecated.